requirements.txt
pyshark>=0.4.5
pycryptodomex>=3.15.0
